# DOCS/api/openapi.yaml
openapi: 3.1.0
info:
  title: CRM FreeLux API
  description: |
    API REST para el sistema CRM de Aceros y Perfiles FreeLux.

    ## Autenticación
    La API usa JWT Bearer tokens. Obtén un token en `POST /api/v1/auth/login`
    y envíalo en el header `Authorization: Bearer <token>`.

    ## Roles
    | Rol | Descripción |
    |-----|-------------|
    | `ADMIN` | Acceso total al sistema |
    | `GERENTE` | Aprobación de precios, reportes, gestión de usuarios |
    | `VENDEDOR` | Crear cotizaciones, ver clientes y productos |
    | `COMPRAS` | Gestión de proveedores, órdenes de compra |
    | `ALMACEN` | Control de inventario, entradas/salidas |

  version: 1.0.0
  contact:
    name: FreeLux CRM
    email: dev@freelux.mx
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Desarrollo local
  - url: https://crm.freelux.mx
    description: Staging / Producción
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Token JWT obtenido de POST /api/v1/auth/login"
  schemas:
...
    UserRole:
      type: string
      enum: [ADMIN, GERENTE, VENDEDOR, COMPRAS, ALMACEN]
      example: VENDEDOR
    QuoteStatus:
      type: string
      enum: [BORRADOR, ENVIADA, ACEPTADA, RECHAZADA, VENCIDA]
      example: ENVIADA
    SaleStatus:
      type: string
      enum: [CONFIRMADA, EN_PREPARACION, LISTA, ENTREGADA, CANCELADA]
      example: CONFIRMADA
...
    UserResponse:
      type: object
      properties:
        id: { type: integer, example: 3 }
        email: { type: string, example: "juan.garcia@freelux.mx" }
        full_name: { type: string, example: "Juan García López" }
        role: { $ref: '#/components/schemas/UserRole' }
        is_active: { type: boolean, example: true }
        created_at: { type: string, format: date-time }

    # ── Clients ────────────────────────────────────────────────
    ClientCreate:
      type: object
      required: [company_name, rfc]
      properties:
        company_name:
          type: string
          example: "IIMSA Infraestructura Industrial SA de CV"
        rfc:
          type: string
          pattern: '^[A-ZÑ&]{3,4}[0-9]{6}[A-Z0-9]{3}$'
          example: "III880415AB7"
        contact_name:
          type: string
          example: "Ing. Roberto Flores"
        phone:
          type: string
          example: "+52 55 1234 5678"
        email:
          type: string
          format: email
          example: "compras@iimsa.com.mx"
        address:
          type: string
          example: "Blvd. Manuel Ávila Camacho 32, Polanco, CDMX"
        credit_limit:
          type: number
          format: float
          example: 500000.00
    ClientResponse:
      allOf:
        - $ref: '#/components/schemas/ClientCreate'
        - type: object
          properties:
            id: { type: integer, example: 42 }
            created_at: { type: string, format: date-time }
            total_purchased_ytd: { type: number, example: 284500.00 }
            last_purchase_date: { type: string, format: date }
            quotes_count: { type: integer, example: 12 }

    # ── Products ───────────────────────────────────────────────
    ProductCreate:
      type: object
      required: [sku, name, unit, cost_price, list_price]
      properties:
        sku:
          type: string
          example: "PTR-4X4-14-6M"
        name:
          type: string
          example: "PTR 4\"x4\"x6m Calibre 14"
        description:
          type: string
          example: "Tubo estructural rectangular 4x4 pulgadas, espesor calibre 14 (1.9mm), longitud 6 metros, acero A36"
        category:
          type: string
          enum: [PERFIL, LAMINA, TUBERIA, PTR, HERRAJE, MALLA, VARILLA, OTRO]
          example: PTR
        unit:
          type: string
          example: "PZA"
        cost_price:
          type: number
          format: float
          example: 820.00
        list_price:
          type: number
          format: float
          example: 980.00
        stock_min:
          type: integer
          example: 20
        stock_max:
          type: integer
          example: 200
        weight_kg:
          type: number
          example: 38.5
        fulfillment_type:
          type: string
          enum: [STOCK, MTO]
          example: STOCK
          description: "STOCK = disponible en almacén, MTO = Make To Order (se pide al recibir venta)"
    ProductResponse:
      allOf:
        - $ref: '#/components/schemas/ProductCreate'
        - type: object
          properties:
            id: { type: integer, example: 15 }
            current_stock: { type: integer, example: 45 }
            stock_status:
              type: string
              enum: [OK, LOW, OUT]
              example: LOW
            created_at: { type: string, format: date-time }

    # ── Suppliers ──────────────────────────────────────────────
    SupplierCreate:
      type: object
      required: [company_name, rfc]
      properties:
        company_name:
          type: string
          example: "Aceros del Norte SA de CV"
        rfc:
          type: string
          example: "AND850312KJ4"
        contact_name:
          type: string
          example: "Ing. Carlos Mendoza"
        phone:
          type: string
          example: "+52 81 8765 4321"
        email:
          type: string
          example: "ventas@acerosdelnorte.com.mx"
        city:
          type: string
          example: "Monterrey, NL"
        payment_terms:
          type: string
          example: "30 días neto"
        lead_time_days:
          type: integer
          example: 3
    SupplierResponse:
      allOf:
        - $ref: '#/components/schemas/SupplierCreate'
        - type: object
          properties:
            id: { type: integer, example: 7 }
            rating: { type: number, example: 4.8 }
            on_time_delivery_rate: { type: number, example: 0.97 }
            total_orders: { type: integer, example: 24 }

    # ── Quotes ─────────────────────────────────────────────────
    QuoteItemCreate:
      type: object
      required: [product_id, quantity, unit_price]
      properties:
        product_id: { type: integer, example: 15 }
        quantity: { type: number, example: 50 }
        unit_price: { type: number, example: 900.00 }
        discount_pct: { type: number, example: 5.0, description: "Porcentaje de descuento 0–100" }
        notes: { type: string, example: "Precio especial por volumen" }
    QuoteCreate:
      type: object
      required: [client_id, items]
      properties:
        client_id: { type: integer, example: 42 }
        items:
          type: array
          items: { $ref: '#/components/schemas/QuoteItemCreate' }
          minItems: 1
        notes:
          type: string
          example: "Entrega en obra del cliente. Incluye flete zona metropolitana."
        validity_days:
          type: integer
          default: 15
          example: 15
        currency:
          type: string
          default: MXN
          example: MXN
    QuoteResponse:
      type: object
      properties:
        id: { type: integer, example: 142 }
        folio: { type: string, example: "COT-2024-0142" }
        client: { $ref: '#/components/schemas/ClientResponse' }
        items:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              product: { $ref: '#/components/schemas/ProductResponse' }
              quantity: { type: number }
              unit_price: { type: number }
              discount_pct: { type: number }
              subtotal: { type: number }
        subtotal: { type: number, example: 117500.00 }
        tax_amount: { type: number, example: 18800.00 }
        total: { type: number, example: 136300.00 }
        currency: { type: string, example: "MXN" }
        status: { $ref: '#/components/schemas/QuoteStatus' }
        requires_approval: { type: boolean, example: false }
        notes: { type: string }
        valid_until: { type: string, format: date }
        sent_at: { type: string, format: date-time, nullable: true }
        created_by: { $ref: '#/components/schemas/UserResponse' }
        created_at: { type: string, format: date-time }
        version: { type: integer, example: 1 }

    # ── Sales ──────────────────────────────────────────────────
    SaleResponse:
      type: object
      properties:
        id: { type: integer, example: 89 }
        folio: { type: string, example: "VTA-2024-0089" }
        remission_folio: { type: string, example: "REM-2024-0089" }
        quote_id: { type: integer, example: 139 }
        client: { $ref: '#/components/schemas/ClientResponse' }
        status: { $ref: '#/components/schemas/SaleStatus' }
        total: { type: number, example: 136300.00 }
        billing_requested: { type: boolean, example: false }
        delivered_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }

    # ── Inventory ──────────────────────────────────────────────
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/clients/{client_id}/sales:
    get:
      tags: [Clients]
      summary: Historial de ventas del cliente
      parameters:
        - name: client_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Historial de ventas
          content:
            application/json:
              example:
                total_ytd: 284500.00
                avg_ticket: 47416.67
                sales_count: 6
                sales: []

  # ── Products ───────────────────────────────────────────────
...
  /api/v1/suppliers:
    get:
      tags: [Suppliers]
      summary: Listar proveedores
      responses:
        '200':
          description: Lista de proveedores
    post:
      tags: [Suppliers]
      summary: Crear proveedor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SupplierCreate' }
      responses:
        '201':
          description: Proveedor creado
  /api/v1/suppliers/{supplier_id}:
    get:
      tags: [Suppliers]
      summary: Obtener proveedor
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Proveedor encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupplierResponse' }
    put:
      tags: [Suppliers]
      summary: Actualizar proveedor
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SupplierCreate' }
      responses:
        '200':
          description: Proveedor actualizado

  # ── Quotes ─────────────────────────────────────────────────
  /api/v1/quotes:
    get:
      tags: [Quotes]
      summary: Listar cotizaciones
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/QuoteStatus' }
        - name: client_id
          in: query
          schema: { type: integer }
        - name: created_by
          in: query
          schema: { type: integer }
        - name: from_date
          in: query
          schema: { type: string, format: date, example: "2024-11-01" }
        - name: to_date
          in: query
          schema: { type: string, format: date, example: "2024-11-30" }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Lista de cotizaciones
    post:
      tags: [Quotes]
      summary: Crear cotización
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuoteCreate' }
            example:
              client_id: 42
              items:
                - product_id: 15
                  quantity: 50
                  unit_price: 900.00
                  notes: "Precio especial por volumen"
                - product_id: 22
                  quantity: 20
                  unit_price: 1250.00
              notes: "Entrega en obra del cliente. Incluye flete ZMCM."
              validity_days: 15
      responses:
        '201':
          description: Cotización creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuoteResponse' }
...
  /api/v1/quotes/{quote_id}:
    get:
      tags: [Quotes]
      summary: Obtener cotización
      parameters:
        - name: quote_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Cotización encontrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuoteResponse' }
    put:
      tags: [Quotes]
      summary: Actualizar cotización (crea nueva versión si ya fue enviada)
      parameters:
        - name: quote_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuoteCreate' }
      responses:
        '200':
          description: Cotización actualizada
  /api/v1/quotes/{quote_id}/pdf:
    get:
      tags: [Quotes]
      summary: Generar y descargar PDF de cotización
      parameters:
        - name: quote_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: PDF de la cotización
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /api/v1/quotes/{quote_id}/send-whatsapp:
    post:
      tags: [Quotes]
      summary: Enviar cotización por WhatsApp
      parameters:
        - name: quote_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            example:
              phone: "+52 55 9876 5432"
              message: "Hola Ing. Flores, adjunto su cotización COT-2024-0142 por $136,300. Válida 15 días."
      responses:
        '202':
          description: Envío programado
          content:
            application/json:
              example: { task_id: "abc123", message: "Mensaje en cola de envío" }
...
  /api/v1/sales:
    get:
      tags: [Sales]
      summary: Listar ventas
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/SaleStatus' }
        - name: client_id
          in: query
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        '200':
          description: Lista de ventas
  /api/v1/sales/{sale_id}:
    get:
      tags: [Sales]
      summary: Obtener venta
      parameters:
        - name: sale_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Venta encontrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SaleResponse' }
...